:recycle: Extract pipeline_accessor.go from pipeline.go
